requirements.txt
streamlit==1.32.0
langchain==0.1.12
langchain-openai==0.0.8
python-dotenv==1.0.1
beautifulsoup4==4.12.3
requests==2.31.0
duckduckgo-search==4.4.3
matplotlib==3.8.3
pandas==2.2.1
google-search-results==2.4.2 